third_party/blink/web_tests/external/wpt/web-nfc/idlharness.https.window-expected.txt

This is a testharness.js-based test.
Found 50 tests; 34 PASS, 16 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS idl_test setup
PASS NFCWriter interface: existence and properties of interface object
PASS NFCWriter interface object length
PASS NFCWriter interface object name
PASS NFCWriter interface: existence and properties of interface prototype object
PASS NFCWriter interface: existence and properties of interface prototype object's "constructor" property
PASS NFCWriter interface: existence and properties of interface prototype object's @@unscopables property
PASS NFCWriter interface: operation push(NDEFMessageSource, NFCPushOptions)
PASS NFCWriter must be primary interface of new NFCWriter();
PASS Stringification of new NFCWriter();
PASS NFCWriter interface: new NFCWriter(); must inherit property "push(NDEFMessageSource, NFCPushOptions)" with the proper type
PASS NFCWriter interface: calling push(NDEFMessageSource, NFCPushOptions) on new NFCWriter(); with too few arguments must throw TypeError
FAIL NFCReader interface: existence and properties of interface object assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface object length assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface object name assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: existence and properties of interface prototype object assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: attribute onreading assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: attribute onerror assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: operation start() assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader interface: operation stop() assert_own_property: self does not have own property "NFCReader" expected property "NFCReader" missing
FAIL NFCReader must be primary interface of new NFCReader(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL Stringification of new NFCReader(); assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "onreading" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "onerror" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "start()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
FAIL NFCReader interface: new NFCReader(); must inherit property "stop()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: NFCReader is not defined"
PASS NFCReadingEvent interface: existence and properties of interface object
PASS NFCReadingEvent interface object length
PASS NFCReadingEvent interface object name
PASS NFCReadingEvent interface: existence and properties of interface prototype object
PASS NFCReadingEvent interface: existence and properties of interface prototype object's "constructor" property
PASS NFCReadingEvent interface: existence and properties of interface prototype object's @@unscopables property
PASS NFCReadingEvent interface: attribute serialNumber
PASS NFCReadingEvent interface: attribute message
PASS NFCReadingEvent must be primary interface of new NFCReadingEvent("reading", { message: {"url":"/custom/path","records":[{"recordType":"text","data":"Hello World"}]} })
PASS Stringification of new NFCReadingEvent("reading", { message: {"url":"/custom/path","records":[{"recordType":"text","data":"Hello World"}]} })
PASS NFCReadingEvent interface: new NFCReadingEvent("reading", { message: {"url":"/custom/path","records":[{"recordType":"text","data":"Hello World"}]} }) must inherit property "serialNumber" with the proper type
PASS NFCReadingEvent interface: new NFCReadingEvent("reading", { message: {"url":"/custom/path","records":[{"recordType":"text","data":"Hello World"}]} }) must inherit property "message" with the proper type
PASS NFCErrorEvent interface: existence and properties of interface object
PASS NFCErrorEvent interface object length
PASS NFCErrorEvent interface object name
PASS NFCErrorEvent interface: existence and properties of interface prototype object
PASS NFCErrorEvent interface: existence and properties of interface prototype object's "constructor" property
PASS NFCErrorEvent interface: existence and properties of interface prototype object's @@unscopables property
PASS NFCErrorEvent interface: attribute error
PASS NFCErrorEvent must be primary interface of new NFCErrorEvent("error", { error: new DOMException() });
PASS Stringification of new NFCErrorEvent("error", { error: new DOMException() });
PASS NFCErrorEvent interface: new NFCErrorEvent("error", { error: new DOMException() }); must inherit property "error" with the proper type
Harness: the test ran to completion.
